requirements.txt
streamlit==1.33.0
pandas==2.2.2
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.1
plotly==5.20.0